requirements.txt
basemap==1.4.1
basemap_data==1.3.2
h5py==3.9.0
matplotlib==3.7.5
numpy==1.24.4
nvidia_modulus==0.1.0
nvidia_modulus.sym==1.0.0
scipy==1.10.1
sympy==1.5.1
torch==2.0.1
tqdm==4.66.1
wandb==0.15.5
tensorboard==2.13.0
pandas==2.0.3